@@ -10,23 +10,24 @@
     "preview": "astro preview"
   },
   "dependencies": {
-    "@astrojs/preact": "^1.2.0",
-    "astro": "^2.0.2",
-    "date-fns": "^2.29.3",
-    "preact": "^10.11.3",
-    "preact-render-to-string": "^5.2.6"
+    "@astrojs/image": "^0.16.9",
+    "@astrojs/preact": "^2.2.0",
+    "astro": "^2.5.5",
+    "preact": "^10.15.1",
+    "preact-render-to-string": "^6.0.3"
   },
   "devDependencies": {
-    "@astrojs/tailwind": "^3.0.0",
-    "@types/node": "^18.11.18",
-    "autoprefixer": "^10.4.13",
-    "critters": "^0.0.16",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.3",
-    "prettier-plugin-astro": "^0.7.2",
-    "prettier-plugin-tailwindcss": "^0.2.2",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.4"
+    "@astrojs/tailwind": "^3.1.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/node": "^18.16.16",
+    "autoprefixer": "^10.4.14",
+    "critters": "^0.0.17",
+    "postcss": "^8.4.24",
+    "prettier": "^2.8.8",
+    "prettier-plugin-astro": "^0.9.1",
+    "prettier-plugin-tailwindcss": "^0.3.0",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.0.4"
   },
   "module": "true"
 }
