@@ -1,4 +1,7 @@
-module.exports = {
+import typography from "@tailwindcss/typography";
+
+/** @type {import('tailwindcss').Config} */
+export default {
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx,astro}",
     "./src/components/**/*.{js,ts,jsx,tsx,astro}",
@@ -6,5 +9,5 @@ module.exports = {
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [typography],
 };
