@@ -16,7 +16,8 @@ const cvWorkCollection = defineCollection({
     location: z.string(),
     from: z.date(),
     to: z.date().or(z.enum(["now"])),
-    type: z.enum(["project", "permanent", "side"]),
+    type: z.enum(["main", "side"]),
+    projectType: z.enum(["permanent", "project"]),
     tasks: z.array(z.string()),
     technologies: z.array(z.string()).or(z.any()),
     contentType: z.enum(["work", "education"]),
